use unsigned int for price and signed int for profit loss

+++ trade-engine/feed/fakefeed/provider.go
@@ -29,7 +29,7 @@ func (p *Provider) RetrievePrices() (map[string]Quote, error) {
 		p.tickers[symbol] = lastPrice
 
 		quotes[symbol] = Quote{
-			LastTradePrice: lastPrice,
+			LastTradePrice: uint64(lastPrice),
 		}
 	}
